Fixing register namespace
