+++ bin/net/erbros/Lottery/payment/methods/.gitignore
@@ -1,0 +1,1 @@
+*.class
