+++ bin/net/erbros/Lottery/register/payment/.gitignore
@@ -1,0 +1,5 @@
+
+*.class
+*.class
+*.class
+*.class
