+++ src/net/erbros/Lottery/PluginListener.java
@@ -1,10 +1,11 @@
 package net.erbros.Lottery;
 
+import net.erbros.Lottery.register.payment.Methods;
+
 import org.bukkit.event.server.PluginDisableEvent;
 import org.bukkit.event.server.PluginEnableEvent;
 import org.bukkit.event.server.ServerListener;
 
-import com.nijikokun.register.payment.Methods;
 
 public class PluginListener extends ServerListener {
     private Lottery plugin;
@@ -22,7 +23,7 @@ public class PluginListener extends ServerListener {
 
             if(check) {
                 this.plugin.Method = null;
-                System.out.println("[" + plugin.info.getName() + "] Payment method was disabled. No longer accepting payments.");
+                System.out.println("[Lottery] Payment method was disabled. No longer accepting payments.");
             }
         }
     }
@@ -34,7 +35,7 @@ public class PluginListener extends ServerListener {
                 // You might want to make this a public variable inside your MAIN class public Method Method = null;
                 // then reference it through this.plugin.Method so that way you can use it in the rest of your plugin ;)
                 this.plugin.Method = this.Methods.getMethod();
-                System.out.println("[" + plugin.info.getName() + "] Payment method found (" + this.plugin.Method.getName() + " version: " + this.plugin.Method.getVersion() + ")");
+                System.out.println("[Lottery] Payment method found (" + this.plugin.Method.getName() + " version: " + this.plugin.Method.getVersion() + ")");
             }
         }
     }
