+++ src/net/erbros/Lottery/payment/methods/MCUR.java
@@ -1,9 +1,9 @@
-package com.nijikokun.register.payment.methods;
+package net.erbros.Lottery.payment.methods;
 
-import com.nijikokun.register.payment.Method;
 
 import me.ashtheking.currency.Currency;
 import me.ashtheking.currency.CurrencyList;
+import net.erbros.Lottery.register.payment.Method;
 
 import org.bukkit.plugin.Plugin;
