+++ src/net/erbros/Lottery/payment/methods/iCo4.java
@@ -1,22 +1,22 @@
-package com.nijikokun.register.payment.methods;
+package net.erbros.Lottery.payment.methods;
 
-import com.iCo6.iConomy;
-import com.iCo6.system.Account;
-import com.iCo6.system.Accounts;
-import com.iCo6.system.Holdings;
+import com.nijiko.coelho.iConomy.iConomy;
+import com.nijiko.coelho.iConomy.system.Account;
 
-import com.nijikokun.register.payment.Method;
+
+import net.erbros.Lottery.register.payment.Method;
 
 import org.bukkit.plugin.Plugin;
 
 /**
- * iConomy 6 Implementation of Method
+ * iConomy 4 Implementation of Method
  *
  * @author Nijikokun <nijikokun@shortmail.com> (@nijikokun)
  * @copyright (c) 2011
  * @license AOL license <http://aol.nexua.org>
  */
-public class iCo6 implements Method {
+@SuppressWarnings("static-access")
+public class iCo4 implements Method {
     private iConomy iConomy;
 
     public iConomy getPlugin() {
@@ -28,11 +28,11 @@ public class iCo6 implements Method {
     }
 
     public String getVersion() {
-        return "6";
+        return "4";
     }
 
     public String format(double amount) {
-        return this.iConomy.format(amount);
+        return this.iConomy.getBank().format(amount);
     }
 
     public boolean hasBanks() {
@@ -44,7 +44,7 @@ public class iCo6 implements Method {
     }
 
     public boolean hasAccount(String name) {
-        return (new Accounts()).exists(name);
+        return this.iConomy.getBank().hasAccount(name);
     }
 
     public boolean hasBankAccount(String bank, String name) {
@@ -52,31 +52,26 @@ public class iCo6 implements Method {
     }
 
     public MethodAccount getAccount(String name) {
-        return new iCoAccount((new Accounts()).get(name));
+        return new iCoAccount(this.iConomy.getBank().getAccount(name));
     }
 
     public MethodBankAccount getBankAccount(String bank, String name) {
         return null;
     }
-
+	
     public boolean isCompatible(Plugin plugin) {
-        try { Class.forName("com.iCo6.IO"); }
-        catch(Exception e) { return false; }
-
-        return plugin.getDescription().getName().equalsIgnoreCase("iconomy") && plugin.getClass().getName().equals("com.iCo6.iConomy") && plugin instanceof iConomy;
+        return plugin.getDescription().getName().equalsIgnoreCase("iconomy") && plugin.getClass().getName().equals("com.nijiko.coelho.iConomy.iConomy") && plugin instanceof iConomy;
     }
 
     public void setPlugin(Plugin plugin) {
         iConomy = (iConomy)plugin;
     }
-
+	
     public class iCoAccount implements MethodAccount {
         private Account account;
-        private Holdings holdings;
 
         public iCoAccount(Account account) {
             this.account = account;
-            this.holdings = account.getHoldings();
         }
 
         public Account getiCoAccount() {
@@ -84,53 +79,53 @@ public class iCo6 implements Method {
         }
 
         public double balance() {
-            return this.holdings.getBalance();
+            return this.account.getBalance();
         }
 
         public boolean set(double amount) {
-            if(this.holdings == null) return false;
-            this.holdings.setBalance(amount);
+            if(this.account == null) return false;
+            this.account.setBalance(amount);
             return true;
         }
 
         public boolean add(double amount) {
-            if(this.holdings == null) return false;
-            this.holdings.add(amount);
+            if(this.account == null) return false;
+            this.account.add(amount);
             return true;
         }
 
         public boolean subtract(double amount) {
-            if(this.holdings == null) return false;
-            this.holdings.subtract(amount);
+            if(this.account == null) return false;
+            this.account.subtract(amount);
             return true;
         }
 
         public boolean multiply(double amount) {
-            if(this.holdings == null) return false;
-            this.holdings.multiply(amount);
+            if(this.account == null) return false;
+            this.account.multiply(amount);
             return true;
         }
 
         public boolean divide(double amount) {
-            if(this.holdings == null) return false;
-            this.holdings.divide(amount);
+            if(this.account == null) return false;
+            this.account.divide(amount);
             return true;
         }
 
         public boolean hasEnough(double amount) {
-            return this.holdings.hasEnough(amount);
+            return this.account.hasEnough(amount);
         }
 
         public boolean hasOver(double amount) {
-            return this.holdings.hasOver(amount);
+            return this.account.hasOver(amount);
         }
 
         public boolean hasUnder(double amount) {
-            return this.holdings.hasUnder(amount);
+            return (this.balance() < amount);
         }
 
         public boolean isNegative() {
-            return this.holdings.isNegative();
+            return this.account.isNegative();
         }
 
         public boolean remove() {
