+++ src/net/erbros/Lottery/payment/methods/iCo6.java
@@ -1,20 +1,24 @@
-package com.nijikokun.register.payment.methods;
+package net.erbros.Lottery.payment.methods;
 
-import com.nijiko.coelho.iConomy.iConomy;
-import com.nijiko.coelho.iConomy.system.Account;
+import com.iCo6.iConomy;
+import com.iCo6.system.Account;
+import com.iCo6.system.Accounts;
+import com.iCo6.system.Holdings;
 
-import com.nijikokun.register.payment.Method;
+
+import net.erbros.Lottery.register.payment.Method;
 
 import org.bukkit.plugin.Plugin;
 
 /**
- * iConomy 4 Implementation of Method
+ * iConomy 6 Implementation of Method
  *
  * @author Nijikokun <nijikokun@shortmail.com> (@nijikokun)
  * @copyright (c) 2011
  * @license AOL license <http://aol.nexua.org>
  */
-public class iCo4 implements Method {
+@SuppressWarnings("static-access")
+public class iCo6 implements Method {
     private iConomy iConomy;
 
     public iConomy getPlugin() {
@@ -26,11 +30,11 @@ public class iCo4 implements Method {
     }
 
     public String getVersion() {
-        return "4";
+        return "6";
     }
 
     public String format(double amount) {
-        return this.iConomy.getBank().format(amount);
+        return this.iConomy.format(amount);
     }
 
     public boolean hasBanks() {
@@ -42,7 +46,7 @@ public class iCo4 implements Method {
     }
 
     public boolean hasAccount(String name) {
-        return this.iConomy.getBank().hasAccount(name);
+        return (new Accounts()).exists(name);
     }
 
     public boolean hasBankAccount(String bank, String name) {
@@ -50,26 +54,31 @@ public class iCo4 implements Method {
     }
 
     public MethodAccount getAccount(String name) {
-        return new iCoAccount(this.iConomy.getBank().getAccount(name));
+        return new iCoAccount((new Accounts()).get(name));
     }
 
     public MethodBankAccount getBankAccount(String bank, String name) {
         return null;
     }
-	
+
     public boolean isCompatible(Plugin plugin) {
-        return plugin.getDescription().getName().equalsIgnoreCase("iconomy") && plugin.getClass().getName().equals("com.nijiko.coelho.iConomy.iConomy") && plugin instanceof iConomy;
+        try { Class.forName("com.iCo6.IO"); }
+        catch(Exception e) { return false; }
+
+        return plugin.getDescription().getName().equalsIgnoreCase("iconomy") && plugin.getClass().getName().equals("com.iCo6.iConomy") && plugin instanceof iConomy;
     }
 
     public void setPlugin(Plugin plugin) {
         iConomy = (iConomy)plugin;
     }
-	
+
     public class iCoAccount implements MethodAccount {
         private Account account;
+        private Holdings holdings;
 
         public iCoAccount(Account account) {
             this.account = account;
+            this.holdings = account.getHoldings();
         }
 
         public Account getiCoAccount() {
@@ -77,53 +86,53 @@ public class iCo4 implements Method {
         }
 
         public double balance() {
-            return this.account.getBalance();
+            return this.holdings.getBalance();
         }
 
         public boolean set(double amount) {
-            if(this.account == null) return false;
-            this.account.setBalance(amount);
+            if(this.holdings == null) return false;
+            this.holdings.setBalance(amount);
             return true;
         }
 
         public boolean add(double amount) {
-            if(this.account == null) return false;
-            this.account.add(amount);
+            if(this.holdings == null) return false;
+            this.holdings.add(amount);
             return true;
         }
 
         public boolean subtract(double amount) {
-            if(this.account == null) return false;
-            this.account.subtract(amount);
+            if(this.holdings == null) return false;
+            this.holdings.subtract(amount);
             return true;
         }
 
         public boolean multiply(double amount) {
-            if(this.account == null) return false;
-            this.account.multiply(amount);
+            if(this.holdings == null) return false;
+            this.holdings.multiply(amount);
             return true;
         }
 
         public boolean divide(double amount) {
-            if(this.account == null) return false;
-            this.account.divide(amount);
+            if(this.holdings == null) return false;
+            this.holdings.divide(amount);
             return true;
         }
 
         public boolean hasEnough(double amount) {
-            return this.account.hasEnough(amount);
+            return this.holdings.hasEnough(amount);
         }
 
         public boolean hasOver(double amount) {
-            return this.account.hasOver(amount);
+            return this.holdings.hasOver(amount);
         }
 
         public boolean hasUnder(double amount) {
-            return (this.balance() < amount);
+            return this.holdings.hasUnder(amount);
         }
 
         public boolean isNegative() {
-            return this.account.isNegative();
+            return this.holdings.isNegative();
         }
 
         public boolean remove() {
